@@ -1,5 +1,8 @@
-const bag              = require('bagofcli');
-const SwaggyJenkinsCli = require('./swaggyjenkinscli');
+"use strict"
+import bag from 'bagofcli';
+import SwaggyJenkinsCli from './swaggyjenkinscli.js';
+
+const DIRNAME = p.dirname(import.meta.url).replace('file://', '');
 
 function _responseToDefinition(inputPath) {
 
@@ -20,13 +23,19 @@ function _responseToDefinition(inputPath) {
  */
 function exec() {
 
-  var actions = {
+  const actions = {
     commands: {
       response2definition: { action: _responseToDefinition }
     }
   };
 
-  bag.command(__dirname, actions);
+  bag.command(DIRNAME, actions);
 }
 
-exports.exec = exec;
+const exports = {
+  exec: exec
+};
+
+export {
+  exports as default
+};
